Add ping check to upload script

+++ _scripts/upload.sh
@@ -1,5 +1,7 @@
 #! /bin/bash
 
+HOST="jamoorev.com"
+
 function getBranch() {
 	echo `echo ${TRAVIS_BRANCH} | sed s/[\/]/-/ `
 }
@@ -21,7 +23,14 @@ if [ "${TRAVIS_REPO_SLUG}" != "JamOORev/Website" ]; then
 	exit 0
 fi
 
-sudo apt-get install ncftp --yes
-cd _site/ && rm -f Gemfile* && ncftpput -R -DD -v -m -u $FTP_USER -p $FTP_PASSWORD jamoorev.com "`getFolder`/" ./* ./.[a-z]*
+#sudo apt-get install ncftp --yes ## Now Handled By Travis In .travis.yml
+
+if ping -qc 20 $HOST >/dev/null; then
+    echo "Host $HOST is up, uploading files"
+    cd _site/ && rm -f Gemfile* && ncftpput -R -DD -v -m -u $FTP_USER -p $FTP_PASSWORD jamoorev.com "`getFolder`/" ./* ./.[a-z]*
+else
+    echo "Host $HOST_TO_CHECK is down, try rebuilding later?"
+    exit 1
+fi
 
 #curl --ftp-create-dirs -T _site/ -u $FTP_USER:$FTP_PASSWORD ftp://jamoorev.com/`getFolder`
